FIXME: Rename config.h to cconfig.h
The global config.h is included on the command line with the gcc '-include'
switch which searches the local directory before the include path. This leads
to a namespace collision with 'src/common/config/config.h'.
