+++ src/bin/lttng-sessiond/load-session-thread.c
@@ -17,7 +17,7 @@
 
 #define _LGPL_SOURCE
 #include <common/error.h>
-#include <common/config/config.h>
+#include <common/config/cconfig.h>
 
 #include "load-session-thread.h"
 #include "lttng-sessiond.h"
