+++ src/bin/lttng-sessiond/save.c
@@ -24,7 +24,7 @@
 
 #include <common/defaults.h>
 #include <common/error.h>
-#include <common/config/config.h>
+#include <common/config/cconfig.h>
 #include <common/utils.h>
 #include <common/runas.h>
 #include <lttng/save-internal.h>
